feat: Guard search queries against common mistakes
