+++ coderd/searchquery/search.go
@@ -6,13 +6,61 @@ import (
 	"strings"
 	"time"
 
-	"github.com/coder/coder/coderd/httpapi"
+	"github.com/google/uuid"
+
 	"golang.org/x/xerrors"
 
 	"github.com/coder/coder/coderd/database"
-
+	"github.com/coder/coder/coderd/httpapi"
 	"github.com/coder/coder/codersdk"
 )
+
+func Audit(query string) (database.GetAuditLogsOffsetParams, []codersdk.ValidationError) {
+	// Always lowercase for all searches.
+	query = strings.ToLower(query)
+	values, errors := searchTerms(query, func(term string, values url.Values) error {
+		values.Set("resource_type", term)
+		return nil
+	})
+	if len(errors) > 0 {
+		return database.GetAuditLogsOffsetParams{}, errors
+	}
+
+	const date_layout = "2006-01-02"
+	parser := httpapi.NewQueryParamParser()
+	filter := database.GetAuditLogsOffsetParams{
+		ResourceID:   parser.UUID(values, uuid.Nil, "resource_id"),
+		Username:     parser.String(values, "", "username"),
+		Email:        parser.String(values, "", "email"),
+		DateFrom:     parser.Time(values, time.Time{}, "date_from", date_layout),
+		DateTo:       parser.Time(values, time.Time{}, "date_to", date_layout),
+		ResourceType: string(httpapi.ParseCustom(parser, values, "", "resource_type", httpapi.ParseEnum[database.ResourceType])),
+		Action:       string(httpapi.ParseCustom(parser, values, "", "action", httpapi.ParseEnum[database.ResourceType])),
+		BuildReason:  string(httpapi.ParseCustom(parser, values, "", "build_reason", httpapi.ParseEnum[database.ResourceType])),
+	}
+	return filter, parser.Errors
+}
+
+func Users(query string) (database.GetUsersParams, []codersdk.ValidationError) {
+	// Always lowercase for all searches.
+	query = strings.ToLower(query)
+	values, errors := searchTerms(query, func(term string, values url.Values) error {
+		values.Set("search", term)
+		return nil
+	})
+	if len(errors) > 0 {
+		return database.GetUsersParams{}, errors
+	}
+
+	parser := httpapi.NewQueryParamParser()
+	filter := database.GetUsersParams{
+		Search:   parser.String(values, "", "search"),
+		Status:   httpapi.ParseCustomList(parser, values, []database.UserStatus{}, "status", httpapi.ParseEnum[database.UserStatus]),
+		RbacRole: parser.Strings(values, []string{}, "role"),
+	}
+	parser.ErrorExcessParams(values)
+	return filter, parser.Errors
+}
 
 func Workspace(query string, page codersdk.Pagination, agentInactiveDisconnectTimeout time.Duration) (database.GetWorkspacesParams, []codersdk.ValidationError) {
 	filter := database.GetWorkspacesParams{
@@ -42,7 +90,7 @@ func Workspace(query string, page codersdk.Pagination, agentInactiveDisconnectTi
 		}
 		return nil
 	})
-	if errors != nil {
+	if len(errors) > 0 {
 		return filter, errors
 	}
 
@@ -86,4 +134,31 @@ func searchTerms(query string, defaultKey func(term string, values url.Values) e
 		}
 	}
 	return searchValues, nil
+}
+
+// splitQueryParameterByDelimiter takes a query string and splits it into the individual elements
+// of the query. Each element is separated by a delimiter. All quoted strings are
+// kept as a single element.
+//
+// Although all our names cannot have spaces, that is a validation error.
+// We should still parse the quoted string as a single value so that validation
+// can properly fail on the space. If we do not, a value of `template:"my name"`
+// will search `template:"my name:name"`, which produces an empty list instead of
+// an error.
+// nolint:revive
+func splitQueryParameterByDelimiter(query string, delimiter rune, maintainQuotes bool) []string {
+	quoted := false
+	parts := strings.FieldsFunc(query, func(r rune) bool {
+		if r == '"' {
+			quoted = !quoted
+		}
+		return !quoted && r == delimiter
+	})
+	if !maintainQuotes {
+		for i, part := range parts {
+			parts[i] = strings.Trim(part, "\"")
+		}
+	}
+
+	return parts
 }
